# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Sistema de Controle Antimicrobianos Veterinários",
    "description": "Arquitetura de um Sistema de Controle de Uso, Vendas e Prescrição de Antimicrobianos Veterinários"
  },
  "security": [{"bearer":[]}],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/status": {
      "get": {
        "summary": "Show all status",
        "tags": [
          "Status"
        ],
        "responses": {
          "200": {
            "description": "Return all status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new status",
        "tags": [
          "Status"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Status/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status created"
          },
          "400": {
            "description": "Error when creating status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an status",
        "tags": [
          "Status"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Status/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status Deleted"
          },
          "400": {
            "description": "Error when deleting status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a status",
        "tags": [
          "Status"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Status/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status Updated"
          },
          "400": {
            "description": "Error when delete status"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/state": {
      "get": {
        "summary": "Show all state",
        "tags": [
          "State"
        ],
        "responses": {
          "200": {
            "description": "Return all state"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new state",
        "tags": [
          "State"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/State/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "state created"
          },
          "400": {
            "description": "Error when creating state"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an state",
        "tags": [
          "State"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/State/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "state Deleted"
          },
          "400": {
            "description": "Error when deleting state"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a state",
        "tags": [
          "State"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/State/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "state Updated"
          },
          "400": {
            "description": "Error when delete state"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/principesActifs": {
      "get": {
        "summary": "Show all principies actifs",
        "tags": [
          "Principies Actifs"
        ],
        "responses": {
          "200": {
            "description": "Return all principies actifs"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new principies actifs",
        "tags": [
          "Principies Actifs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PrincipiesActifs/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principies actifs created"
          },
          "400": {
            "description": "Error when creating principies actifs"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an principies actifs",
        "tags": [
          "Principies Actifs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PrincipiesActifs/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principies actifs Deleted"
          },
          "400": {
            "description": "Error when deleting principies actifs"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a principies actifs",
        "tags": [
          "Principies Actifs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PrincipiesActifs/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Principies actifs updated"
          },
          "400": {
            "description": "Error when delete principies actifs"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/pharmacologicClass": {
      "get": {
        "summary": "Show all pharmacologic class",
        "tags": [
          "Pharmacologic Class"
        ],
        "responses": {
          "200": {
            "description": "Return all pharmacologic class"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new pharmacologic class",
        "tags": [
          "Pharmacologic Class"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PharmacologicClass/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pharmacologic class created"
          },
          "400": {
            "description": "Error when creating pharmacologic class"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an pharmacologic class",
        "tags": [
          "Pharmacologic Class"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PharmacologicClass/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pharmacologic class Deleted"
          },
          "400": {
            "description": "Error when deleting pharmacologic class"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a pharmacologic class",
        "tags": [
          "Pharmacologic Class"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/PharmacologicClass/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pharmacologic class updated"
          },
          "400": {
            "description": "Error when delete pharmacologic class"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/uc": {
      "get": {
        "summary": "Show all Uc",
        "tags": [
          "Uc"
        ],
        "responses": {
          "200": {
            "description": "Return all Uc"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create an new Uc",
        "tags": [
          "Uc"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Uc/Create"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uc created"
          },
          "400": {
            "description": "Error when creating Uc"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "summary": "Update an Uc",
        "tags": [
          "Uc"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Uc/Update"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uc Deleted"
          },
          "400": {
            "description": "Error when deleting Uc"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Error id dont was found"
          }
        }
      },
      "delete": {
        "summary": "Delete a Uc",
        "tags": [
          "Uc"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schema/Uc/Delete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uc updated"
          },
          "400": {
            "description": "Error when delete Uc"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}